backend/evaluations/requirements.txt
fastapi
langchain
langchain-core
langchain-community
pypdf
pymupdf
sentence-transformers
faiss-cpu

langchain-groq
uvicorn
python-dotenv
